Fix hourly summary display
Previous code reset state variables every time an hourly report was shown,
even if it was the hourly report at the end of execution, not on the hour.
This means that the history required to show an accurate hourly report is
broken when the script is running every minute, etc.

So we add a parameter to allow only reseting state when showing a report
on actual hour changes.