Support testing against a plain input file
Don't try to configure the input properties, or write commands unless it
actually looks like we've been asked to open a device. This makes it easier to
test the logic by passing in a plain file containing what we'd expect from a
serial port, instead of a serial port itself.