Process entire buffer after a read
The previous refactor introduced behaviour where we only processed the first
line found in the buffer after each read, subsequent lines were processed after
subsequent reads. Given that the number of reads is almost always less than the
number of lines, this meant the buffer never emptied of most lines. Whoops.

Now we process as many lines as can be found in the buffer before going back
for another read.