Ensure hello is sent regularly
Previously we only sent config on first start (if not in low power), or after a
config update, this left the hole that if you enable low power via a filesystem
upload (pio run -t uploadfs) the MQTT hello/config topic never got updated.

Now we'll always send a hello when we boot up, but if we're in low power mode,
it's delayed until the first report (rather than rigth at boot), to maximise
how long the radio stays off. This is a slight power regression, but should be
OK.

This also resolves the TODO, now we get config check-in while in low-power, and
the callback is setup, so if we're quick we can send config updates/disable
low-power before the next sleep.