Major refactor to add flexible sensor/pin support.
The major improved flexibility comes from two dimensions:
1) Support for different sensor models (e.g. different pulse to L computations)
2) A new config option to dynamically instantiate any supported sensor model
   and map it to a particular pin.

Together it means we no longer have a single sensor type hard-coded on only 2
pins, and instead can instantiate as many sensor instances (of any supported
model) as we like, limited only by the number of pins available to connect to.

Initially in this commit there are 2 models of sensor supported:
* YFB-10 - what the code has always handled
* FS400A - new; same principle but different pulse computation.

As part of the refactor the growing codebase has been split into several files
for improved readability and minor improvements to debug printing and the
status page have also been added.