Use a friendly name for each sensor for reporting
Extend sensorSpec, the Sensor (sub)classes and the smarthouse receiver to
support a friendly name for each configured sensor which is used in the MQTT
reporting and prometheus exports.

The motivation is to both make things easier to understand without having to
remember what's on each pin number, and also enable time series to be preserved
(e.g. no label changes) if/when the pin configuration (or potentially even the
reporting node if names are known to be unique across nodes) changes.